requirements.txt
python-dotenv>=1.0.0
# Install lighter with: pip install git+https://github.com/elliottech/lighter-python.git